Added xAFAttribute resource
DSC resource for AF Attributes. Intended for use with configuration
attributes, like those in the Web API configuration element.

Supports all of the basic attribute value data types, including arrays
of those types. Values must still be specified in DSC configurations as
strings, they will be cast to the intended type when written to the
attribute.